Add input instructions for day 8 challenge
Introduced a new set of input instructions coded in key-value pairs, strings, and sequences of characters for the day 8 puzzle challenge. This large dataset will be utilized for more complicated and diverse decryption and analysis exercises. The goal is to improve the problem-solving skills and enhance the understanding of complex data manipulation.